requirements.txt
requests
python-dotenv
pynacl
base58